# spanconfig/kvaccessor: re-work validation query