The previous validation query used point scans for each span being
upserted, and ended up being very memory intensive. This PR changes the
validation query to instead validate a single large span straddling all
the individual spans being upserted. The benchmark added in the previous
commit demonstrates the speed up:

   dev bench pkg/spanconfig/spanconfigkvaccessor
    -f=BenchmarkKVAccessorUpdate -v --bench-time=20x

Before:

  BenchmarkKVAccessorUpdate/batch-size=100-10       20     61.55 ms/batch     1.625 records/ms
  BenchmarkKVAccessorUpdate/batch-size=1000-10      20      1645 ms/batch     0.6078 records/ms
  BenchmarkKVAccessorUpdate/batch-size=10000-10     ran into memory bound account errors

After:

  BenchmarkKVAccessorUpdate/batch-size=100-10      200     11.25 ms/batch     8.889 records/ms
  BenchmarkKVAccessorUpdate/batch-size=1000-10     200     48.40 ms/batch     20.66 records/ms
  BenchmarkKVAccessorUpdate/batch-size=10000-10    200     222.6 ms/batch     44.92 records/ms

We want to validate that upserting spans does not break the invariant
that spans in the table are non-overlapping. We only need to validate
the spans that are being upserted, and can use a query similar to what
we do in GetSpanConfigRecords. For a single upserted span, we want
effectively validate using:

  -- verify only a single span overlaps with [$start, $end)
  SELECT count(*) = 1 FROM system.span_configurations
   WHERE start_key < $end AND end_key > $start

With the naive form above that translates to an unbounded index scan on
followed by a filter. Since start_key < end_key, and that spans are
non-overlapping, we can instead do the following:

  SELECT bool_and(valid) FROM (
   SELECT bool_and(prev_end_key IS NULL OR start_key >= prev_end_key) AS valid FROM (
    SELECT start_key, lag(end_key, 1) OVER (ORDER BY start_key) AS prev_end_key FROM span_configurations
    WHERE start_key >= $start AND start_key < $end
   )
   UNION ALL
   SELECT * FROM (
    SELECT $start >= end_key FROM span_configurations
    WHERE start_key < $start ORDER BY start_key DESC LIMIT 1
   )
  )

The idea is to first find all spans that start within the span being
upserted[^1], compare each start key to the preceding end key[^2] (if any)
and ensure that they're non-overlapping[^3]. We also verify the span with
the start key immediately preceding the span being upserted[^4]; ensuring
that our upserted span does not overlap with it[^5].

When multiple spans are being upserted, we validate instead the span
straddling all the individual spans being upserted, i.e.
[$smallest-start-key, $largest-end-key).

[^1]: WHERE start_key >= $start AND start_key < $end
[^2]: lag(end_key, 1) OVER (ORDER BY start_key) AS prev_end_key
[^3]: start_key >= prev_end_key
[^4]: WHERE start_key < $start ORDER BY start_key DESC LIMIT 1
[^5]: $start >= end_key

Release note: None
Release justification: low risk, high benefit change